Muhammad Reza Syachrani/samsatinternship/requirements.txt
backports.csv==1.0.7
defusedxml==0.6.0
diff-match-patch==20200713
Django==1.11.29
django-import-export==1.2.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==2.6.4
pathlib==1.0
pytz==2020.5
PyYAML==5.3.1
tablib==0.14.0
xlrd==2.0.1
xlwt==1.3.0
gunicorn==19.0.0
whitenoise==5.2.0
sklearn==0.0
pandas==1.1.5
matplotlib==3.3.3